fix some php8 issues

+++ src/Algorithm/Sorting/RadixSort.php
@@ -65,7 +65,7 @@ class RadixSort implements ISortable {
         $count  = array_fill(0, 10, 0);
 
         for ($i = 0; $i < $n; $i++) {
-            $count[($array[$i] / $exponent) % 10]++;
+            $count[(int) ($array[$i] / $exponent) % 10]++;
         }
 
         for ($i = 1; $i < 10; $i++) {
@@ -73,8 +73,8 @@ class RadixSort implements ISortable {
         }
 
         for ($i = $n - 1; $i >= 0; $i--) {
-            $output[$count[($array[$i] / $exponent) % 10] - 1] = $array[$i];
-            $count[($array[$i] / $exponent) % 10]--;
+            $output[$count[(int) ($array[(int) $i] / $exponent) % 10] - 1] = $array[$i];
+            $count[(int) ($array[(int) $i] / $exponent) % 10]--;
         }
 
         for ($i = 0; $i < $n; $i++) {
